Change behaviour of HandleMessage
Based on feedback from @poveden in #279, the implementation of the Win32
event handling method (ChromaImplementation.HandleMessage) has been
changed to return `false` if event handling has not been registered,
or the handle not having the expected value, as opposed to throwing an
exception. This is to avoid unexpected behaviour in event processing.