Update device info structure
The old boolean "Connected" field in the SDK has been changed
to a DWORD specifying how many devices are connected.
The SdkDeviceInfo struct in Colore has been updated to reflect this change.

The field for the device count is renamed to "ConnectedCount" in Colore
with a new helper property "Connected" to retain the old behaviour.