Detect non-connected devices in QueryDevice
If the SDK returns the ERROR_DEVICE_NOT_CONNECTED error when calling
the QueryDevice API, do not throw a NativeCallException but instead
detect this error and return a device info struct where the connected
count is set to zero.

A new enum member "Unknown" is added to the DeviceType enumeration
to cover this use case.

Fixes #284.